2020-03: usize -> u64

+++ crates/core/src/year2020/day03.rs
@@ -33,7 +33,7 @@ impl Map {
     }
 }
 
-pub fn solve(input: &mut Input) -> Result<usize, String> {
+pub fn solve(input: &mut Input) -> Result<u64, String> {
     let map = Map::parse(input.text)?;
     let slopes = input.part_values(vec![(3, 1)], vec![(1, 1), (3, 1), (5, 1), (7, 1), (1, 2)]);
 
@@ -55,7 +55,7 @@ pub fn solve(input: &mut Input) -> Result<usize, String> {
         .filter(|tree| *tree)
         .count();
 
-        acc * trees_seen_now
+        acc * (trees_seen_now as u64)
     }))
 }
